+++ EssentialFeed/EssentialFeediOS/Composer/FeedUIComposer.swift
@@ -80,33 +80,3 @@ final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
     }
     
 }
-
-final class FeedImageDataLoaderPresentationAdapter<View: FeedImageView, Image>: FeedImageCellControllerDelegate where View.Image == Image {
-    private var task: FeedImageDataLoaderTask?
-    private let model: FeedImage
-    private let imageLoader: FeedImageDataLoader
-
-    var presenter: FeedImagePresenter<View, Image>?
-    
-    init(model: FeedImage, imageLoader: FeedImageDataLoader) {
-        self.model = model
-        self.imageLoader = imageLoader
-    }
-    
-    func didRequestImage() {
-        presenter?.didStartLaoding(for: model)
-        task = imageLoader.loadImageData(from: model.url) { [weak self, model] result in
-            switch result {
-            case let .success(data):
-                self?.presenter?.didFinishDataLoading(with: data, for: model)
-            case let .failure(error):
-                self?.presenter?.didFinishDataLoading(with: error, for: model)
-            }
-        }
-    }
-    
-    func didCancelImageRequest() {
-        task?.cancel()
-    }
-    
-}
